Made changes according to PR comments Replaced e2e-test.groovy with e2e-test-cycloud.groovy Removed e2e-test-cycloud.groovy Remove catch block to return the correct error codes
